Change New-AzureDataFactoryEncryptValue to support on premises Windows Auth encryption. a) Add an optional parameter "WindowsCredential", which will be used in Windows Auth encryption scenario.
b) Upgrade the Gateway Client Nuget package to 1.1.0 version in which the winAuth is supported.

c) Upgrade the help contents (This will be sent to UE review).

d) A new property "Version" in ConnectionInfo object is used, so this change depends on the new version data factory nuget package. Pending the check in until the datafactory nuget package is upgraded.