Use explicit exports in module manifest
When wildcards are used in the module manifest, PowerShell is slower when analyzing modules to see what commands are exported.

This is typically a one time cost (first time you start up), but in a container, it is a cost paid pretty much every time you run PowerShell in the container.

This analysis PowerShell does costs around 250ms, so explicitly declaring the exports is a noticable win.

This change should not cause any change in behavior, it should only impact performance.