Fixes to default threading values and data sources
Default threading values are now set to 5 and 10 for concurrent files
and threads per file, respectively.
Also, improved the experience for listing data sources to list all data
sources and tag them as default if they are the default. Retrieving a
specific data source continues to behave the same as it always has.