Added range for ASN number
Currently we allow people to add an ASN that is out of the defined range for ASNs. 
If they then try and use the Gateway later to add 'Connections' with BGP enabled - they will create but never reach  the Connected state - with no visible errors to the user. 
Adding in this validation will prevent this scenario occurring and save effort on complicated support cases etc.