Update PowerShell to latest package for ADL
This includes a rework of the PowerShell friendly client and smart
settings for file and folder upload/download.
I also re-ran all tests to ensure that they are all still passing, which
resulted in discovering that the MASK is now properly populated
immediately when creating an account, so had to update the permissions
tests.