Addressed some more comments
I switched from long to int due to potential loss of precision when
using GetRange() to return a subset of the collection that is desired.
Additionally, using uint is not recommended since all list management
methods use signed ints. Instead we are limiting the number of rows that
can be specified to: 2147483647, which should be fine for the first
version. If we have customer complaints about needing to preview more
than 2 billion rows, then we can revisit at that time.