Remove email addresses from UPN query.
Before this change, when a UPN argument is passed
to one of the vault policy cmdlets the query
for object id would include email addresses as
well.  After this change UPN will mean UPN and
there will be a seperate email argument if the
user wants to look up the object id by email
address.