Record Auditing tests that use non-classic storage.
Separated auditing and threat detection classic storage tests so that the rest don't need classic storage client credentials to be configured.

Changed AzureEndpointsCommunicator in product code to avoid using SqlManagementClient for direct HTTP requests because it was causing ObjectDisposedException when recording tests. This change also reduces the number of HTTP client objects that are created which can be good for reducing resources used. Also refactored AzureEndpointsCommunicator to avoid using exceptions for control flow.