Updating AzureRM.Logic module page:  deleting deprecated cmdlets
Hi - I removed cmdlets that are no longer part of the module from the index.md - as this causes there to be 404 errors on docs when users try to access these cmdlets from the module landing page.  This fix will remove the "Others" section from this page.  For specifics, please see this page below and click on one of the cmdlets in the "Others" section.

https://docs.microsoft.com/en-us/powershell/module/azurerm.logicapp/?view=azurermps-4.1.0#logic_apps