Suppress breaking changes error messages
- Discussed with PowerShell team and got confirmation from them that the
"The type of property X..." error messages can be safely ignored. They
are not breaking changes since the type was simply renamed and the
properties were preserved.
- For "The property 'Type'..." error messages can also be ignored since
those are discriminatory types used for serialization/deserialization
and the property is no longer produced when using Swagger/AutoRest based
SDKs.
- For the remaining DateTime type error, I do not see any reference to
that property in the Commands.StreamAnalytics.* projects. I believe this
may be a case where the property is not exposed to the customer and
therefore means there is no breaking change associated with this error.