Fix Importing Context File Multiple Times Using -Path
This is a fix for issue 4715.

When you import a context json file, selecting a new subscription
(Select-AzureRmSubscription), and then importing the same context file
again, the context it was on prior to importing the context file a
second time gets invalidated. Doing so causes Select-AzureRmSubscription
to fail and defaults to the default context loaded from the json file.

This is because the DefaultContextKey was not reset in the TryCopyProfile()
method call. This change sets it to the profile it's trying to copy.