Fix Import-AzureRmContext -AzureContext
Calling -AzureContext (Connect-AzureRmAccount) (as shown in the
documentation pages for this cmdlet) throws an exception stating that
you cannot modify a collection... When the cmdlet gets to trying to copy
the ExtendedProperties (in the DefaultContext), the old model and the
model to copy to are the same, and therefore throws the exception
mentioned.

This change checks that the DefaultContext in "other" is not the same as
the one it's trying to copy.

On branch fix-import-azurermcontext