Fix Get-AzureRmTrafficManagerProfile Parameters
Calling Get-AzureRmTrafficManagerProfile with the -Name parameter will
return nothing because the cmdlet requires the -ResourceGroupName
parameter when using -Name.

I've updated the cmdlet parameters to use ParameterSets, which is
similar to how other cmdlets handle this situation. The user can
either call the cmdlet without any parameters, or with the
-ResourceGroupName parameter, or if calling using the -Name parameter,
they will now be prompted to enter in a value for the -ResourceGroupName
parameter.