Azurerm.Profile
Adding Azurerm.profile to list of required modules, which is missing.  This creates an issue where, after the azurerm module has been installed on a fresh machine (did not have the module installed prior), the azurerm.profile module is not being imported into the powershell session with the other azurerm modules, and the user has to either a) import the module manually, or b) restart the powershell console.  By adding the azurerm.profile module to the list of required modules (why was it removed in the first place?), this should fix this issue.