Update Get-AzStorageBlob Blob parameter definition
Existing definition of Blob parameter incorrectly states that it cannot take wildcard characters, but that any blobs matching the provided string will be returned. Actual behaviour I saw was the opposite - if wildcard charactes are provided they are used for string matching purposes, and if no wildcard characters exist in the given string, only blobs with a name matching the given string exactly are returned.
I have a couple screenshots showing this behaviour if they are desired.