Command is not correct
We can use this command for secret not certificates because this command is not correct for certificates of KeyVault. When we use this command, the id will be below. I think this id is not correct for AppGW. 
https://xxxxx.vault.azure.net:443/certificates/xxxx/

The should be below. 

"KeyVaultSecretId": "https://xxxx.vault.azure.net:443/secrets/xxxx/"

When I used below command, I encountered error.

$secretId = $secret.Id.Replace($secret.Version, "")

"Set-AzApplicationGateway : SecretId 'https://xxxx.vault.azure.net:443/certificates/selftest/' specified in '/subscriptions/xxxxxxxxx/resourceGroups/apgw/p
roviders/Microsoft.Network/applicationGateways/AppGW/sslCertificates/SelfKeyCert' is invalid."

When I used below command, the operation was successful. So I think this document is not correct.

$secretId = $certificate.SecretId.Replace($certificate.Version, "")