Adding firewall prerequisite information (#13385)
While trying to follow the example in this page I was getting GatewayTimeout errors. After updating the firewall to allow Azure services the examples work fine. This is important information to the success of the reader executing the example code. As this is essential information required for user success it has been highlighted as Important.