Update Get-AzKeyVaultCertificate.md - Get cert and save it as pfx section to work with PowerShell Core (#13557)
* Update Get-AzKeyVaultCertificate.md

The existing syntax does not seem to work in PowerShell Core.  Using the proposed syntax works with both PowerShell Desktop and Core.  When Calling the 'import' in this line:
$x509Cert.Import($secretByte, "", "Exportable,PersistKeySet") in PowerShell core, an error is received "Import" with "3" argument(s): "X509Certificate is immutable on this platform. Use the equivalent constructor instead."  
Declaring the cert object like this works fine:
$x509Cert = New-Object System.Security.Cryptography.X509Certificates.X509Certificate2($secretByte, "", "Exportable,PersistKeySet")

The recommended change is a combination of the existing MS Documentation and the above which was pulled from: https://gist.github.com/holgerjay/0db1f759f93cba268d671341b787472a

* Remove duplicated line in example

Co-authored-by: Yeming Liu <Yeming.Liu@microsoft.com>