Update Set-AzSqlDatabaseVulnerabilityAssessmentRuleBaseline.md (#14432)
The default powershell array does not parse the values correctly when there is a single array , However it works fine when there are multiple arrays. 
The workaround would be to not use the default array and use [System.Collections.ArrayList] as mentioned in here https://github.com/Azure/azure-powershell/issues/12092.
Please update the documentation , we have customers wondering why this command is not working on single array. 
I have added the script above as well in Example 1.