Update New-AzDataLakeGen2Item.md
From my explorations, it appears that New-AzDataLakeGen2Item has no output, and doesn't throw on unsuccesful transfer when the target file already exists (even with -Force option).

In my opinion it would be better if this Cmdlet threw an exception if no new file was created, and/or if there was an option to overwrite an existing file. However, right now this docs change reflects the actual current behaviour.