Update Az.StackHCI.psm1
Since you already invoke-command to cluster, there is no point to specify cluster name in command. And it also fails if Distributed Domain Name is used for cluster name resource, because then it causes double hop and authentication will fail as it first resolve one ($ClusterNameSession) and later another (-ClusterName ...).