Markup rendering issue?
The sentence "The default value is Block." was being displayed after the last bullet point, which is obviously not right. I added an extra line to get it to be displayed on a new line. I previewed this in an online Markup viewer to make sure it looked OK, since I don't see an option to preview my change here.