{AzureStorage} fixes Azure/azure-powershell#19734
fixes Azure/azure-powershell#19734

Docs link: https://learn.microsoft.com/en-us/powershell/module/az.storage/new-azstoragecontainersastoken?view=azps-8.3.0#-permission

The above documentation for New-AzStorageContainerSASToken doesn't explain what the valid values for -Permission are. All it states is "It is important to note that this is a string, like rwd". This PR addresses the issue and provides the hyperlink to supported permissions.