Sample needs storage context reference (#19782)
* Sample needs storage context reference

This sample is great but customers and myself are always trying to figure out how to populate that context variable. Just updating the sample to show the full sequence

* Fixing context

There are lots of ways to setup a storage context. Using a key here instead of oauth token which is more appropriate for the sample.