container name not supported with upper case (#20012)
TargetContainerName - This name may only contain lowercase letters, numbers, and hyphens, and must begin with a letter or a number. Each hyphen must be preceded and followed by a non-hyphen character. The name must also be between 3 and 63 characters long.

the command "Backup-AzApiManagement" fails if TargetContainerName has value in upper case