{AzureNetworking} fixes Azure/azure-powershell#19436 Support for -WhatIf parameter (#20473)
* {AzureNetworking} fixes Azure/azure-powershell#19436 Support for -WhatIf parameter

fixes #19436 Support for -WhatIf parameter

All 'AzIpGroup' cmdlets that make changes support the -whatIf parameters. New-AzIpGroup and Remove-AzIpGroup honor this, displaying a "What If: " message and not making changes.

Set-AzIpGroup supports -whatif according to the documentation and accepts the parameter, but ignores it and changes the configuration of the resource instead. This PR fixes a Bug for Set-AzIpGroup to support the -whatif parameter.

* Update ChangeLog.md