{AzureNetworking} fixes Azure/azure-powershell#20416 Add support for what-if parameter (#20491)
* {AzureNetworking} fixes Azure/azure-powershell#20416 Add support for what-if parameter

fixes Azure/azure-powershell#20416 Add support for what-if parameter

As per our docs, the Add-AzLoadBalancerFrontendIpConfig, Add-AzLoadBalancerProbeConfig, Add-AzLoadBalancerBackendAddressPoolConfig cmdlets support `what-if` parameter. But currently its not honoring the `what-if` switch. This PR fixes this issue.

* Update AddAzureRmLoadBalancerBackendAddressPoolConfigCommand.cs

fixes #20416 Add support for what-if parameter

As per our docs, the Add-AzLoadBalancerFrontendIpConfig, Add-AzLoadBalancerProbeConfig, Add-AzLoadBalancerBackendAddressPoolConfig cmdlets support what-if parameter. But currently its not honoring the what-if switch. This PR fixes this issue.

* Update AddAzureRmLoadBalancerFrontendIpConfigCommand.cs

fixes #20416 Add support for what-if parameter

As per our docs, the Add-AzLoadBalancerFrontendIpConfig, Add-AzLoadBalancerProbeConfig, Add-AzLoadBalancerBackendAddressPoolConfig cmdlets support what-if parameter. But currently its not honoring the what-if switch. This PR fixes this issue.

* Update ChangeLog.md

* Update RemoveAzureRmLoadBalancerInboundNatRuleConfigCommand.cs

* Update NewAzureRmLoadBalancerRuleConfigCommand.cs

* Update SetAzureRMLoadBalancerCommand.cs

* Update ChangeLog.md

* Update ChangeLog.md