{AzureDataShare} fixes Azure/azure-powershell#20822 fix the description
fixes Azure/azure-powershell#20822 fix the description

The StorageAccountResourceId parameter in the "New-AzDataShareDataSetMapping" PowerShell cmdlet is used to specify the resource ID of the Azure storage account that is the source of the data being shared. This parameter is required when creating a new data set mapping for an Azure Data Share and is used to identify the storage account from which the data will be shared. The resource ID of the storage account can be obtained from the Azure portal or through Azure PowerShell.

The AccountName parameter in the "New-AzDataShareDataSetMapping" PowerShell cmdlet is used to specify the name of the Azure storage account that is the source of the data being shared. This parameter is required when creating a new data set mapping for an Azure Data Share and is used to identify the storage account from which the data will be shared. The account name of the storage account must be unique within Azure and can be used to access the storage account and the data stored within it.

The ResourceGroup parameter in the "New-AzDataShareDataSetMapping" PowerShell cmdlet is used to specify the name of the Azure resource group that contains the Azure storage account being shared. This parameter is required when creating a new data set mapping for an Azure Data Share and is used to identify the resource group that holds the storage account from which the data will be shared. The resource group can be used to manage the resources within it, such as the storage account and the data set mapping.