Clarify behavior of AcountEnabled and Password (#21006)
* Clarify behavior of AcountEnabled and Password

`Update-AzAdUser` in `Az.Resources` v6.0.1 automatically enables the user account when the password changes, ignoring the `-AccountEnabled` parameter. In previous versions of the cmdlet the password could be updated at the same time the account was disabled. That now requires two separate calls to `Update-AzAdUser`: one to set the password, and a second to disable the account.

* Update src/Resources/Resources/ChangeLog.md

---------

Co-authored-by: Beisi Zhou <zhoubeisi@gmail.com>