Updated Example 3 of the Connect-AzAccount.md page to use the correct PowerShell variable.
In line 137 the variable created is $SecurePassword but in line 140 $SecuredPassword was being used. Copying this example leads to the following error: New-Object : Exception calling ".ctor" with "2" argument(s): "Cannot process argument because the value of argument
"password" is null. Change the value of argument "password" to a non-null value."