Refactor of Deployment cmdlets base class. (#24159)
* Refactor of DeploymentCmdletBase.

This refactor is mostly concerning the verbosity of dynamic parameter
logic within the cmdlet. This was broken up into the already existing
TemplateUtility and the newly created ParameterUtility. Documentation
was also updated and other pieces of code broken up. An additional purpose of this
PR was to allow the deployment stack cmdlets to have an easier time
implementing dynamic parameters.

* added changelog addition.

* Removed incorrect checks in existing test cases after changes.