Update Update-AzVmssInstance.md
This change is proposed due to a change request by a customer from a case that I worked on as the Microsoft Azure Support Engineer. The  customer submitted was VMSS update failure case to Azure Support. The case included an error message that stated that there was a "Conflict with concurrent request". The Cx used the parameters in this document to apply the updates to their VMSS instances. After some technical analysis with their cloud solution architect, it was the discovered that the -AsJob parameter in their commands was causing this error since the jobs were being asynchronously executed as background jobs and throws this concurrent request conflict error. I believe it would be good to state this in this documentation to help other Azure users prevent this error from occurring and resulting in an unhealthy VMSS instance. The customer also made this document update request. 

Here is a sample of the command that the customer was running: 

$VMSS = Get-AzVmssVM -ResourceGroupName "myResourceGroup" -VmScaleSetName "myScaleSet"
foreach ($bi in $VMSS)
 {
    Update-AzVmssInstance -ResourceGroupName "myResourceGroup" -VMScaleSetName "myScaleSet" -InstanceId "myInstanceID" -AsJob | Out-Null
}

Thanks!