[Batch] Fix null CommonEnvironmentSettings values
Setting null values inside nested objects such as
CommonEnvironmentSettings throw an 'Object reference not set to an
instance of an object' error.

Allow nulls to pass through to the Batch service so that objects
which support them work properly, and those that don't throw a
more useful error message from the service itself.